[FLAVA]Move image masking transform to core
[ghstack-poisoned]
